fix(blackjack-ui): use long timer values

+++ Intersect.Client.Core/Interface/Game/BlackjackWindow.TableSkin.cs
@@ -144,7 +144,7 @@ internal sealed partial class BlackjackWindow
         UpdateTableSkinTimer(timed && seconds > 0, seconds);
     }
 
-    private void UpdateTableSkinTimer(bool visible, int seconds)
+    private void UpdateTableSkinTimer(bool visible, long seconds)
     {
         SetTableSkinVisible(_timerSkin, visible);
